calc_fusion_vaf: can now specify target gene, window=0 ignores read pairs

+++ calc_fusion_vaf/calc_fusion_vaf.py
@@ -10,7 +10,7 @@ Assumptions
 -----------
 - The breakpoint in the fusion sequences is right in the middle,
     which is what Factera does by default.
-- BWA and samtools should be located in the PATH environment variable.
+- BWA should be located in the PATH environment variable.
 - This was meant for capture-based sequencing data, where one of the
     two genes in a fusion is targeted. This has implications for the
     calculation of the VAF, where the coverage for the wild-type
@@ -19,18 +19,11 @@ Assumptions
 
 Known Issues
 ------------
-- Currently needs a reference genome to be specified. The feature
-    of automatically retrieving the wild-type gene sequences from
-    the Ensembl REST API hasn't been implemented yet. This is
-    partially due to the problem that gene symbols are passed, not
-    stable gene IDs. Therefore, Ensembl might return ambiguous
-    genes. Another option is that the breakpoint position can be
-    considered when unambiguously determining the gene from Ensembl.
-    Also, the breakpoint positions are based on the reference.
 - Currently, reads that span the edges of the window are included.
     I believe it would be better if the reads need to be entirely
     within the window.
-- I do not remove duplicates.
+- Duplicates aren't being removed, because rmdup in samtools 1.1
+    seems to be broken under certain conditions.
 """
 
 __author__ = 'Bruno Grande'
@@ -46,8 +39,9 @@ import subprocess
 import pysam
 from collections import defaultdict
 
+# IMPORTANT! Make sure the fieldnames here correspond to those below when compiling results
 OUTPUT_FIELDNAMES = [
-    'fusion', 'gene_1', 'gene_2', 'breakpoint_1', 'breakpoint_2', 'fusion_spanning_reads',
+    'fusion_id', 'gene_1', 'gene_2', 'breakpoint_1', 'breakpoint_2', 'fusion_spanning_reads',
     'fusion_spanning_read_pairs', 'fusion_total_support', 'wildtype_spanning_reads',
     'wildtype_spanning_read_pairs', 'wildtype_total_support', 'variant_allele_fraction']
 
@@ -55,18 +49,22 @@ OUTPUT_FIELDNAMES = [
 def main():
     # Argument parsing
     parser = argparse.ArgumentParser()
-    parser.add_argument('reference_genome', help='Location of reference genome in FASTA format')
-    parser.add_argument('fastq_files', nargs=2, help='Input FASTQ files')
+    parser.add_argument('reference_genome', help='Location of reference genome in FASTA format.')
+    parser.add_argument('fastq_files', nargs=2, help='Input FASTQ files.')
     parser.add_argument('factera_fusions_txt', type=argparse.FileType('r'),
                         help='Detailed Factera fusions output file (*.fusions.txt).')
     parser.add_argument('--window', '-w', type=int, default=2000,
-                        help='Size of window around breakpoint positions')
+                        help='Size of window around breakpoints. If 0, read pairs are ignored.')
     parser.add_argument('--min_overlap', '-m', type=int, default=10,
                         help='Minimum number of overlapping bases for a spannning read.')
-    parser.add_argument('--output_dir', '-o', default='.', help='Output directory')
-    parser.add_argument('--log', '-l', help='Specify log file. Otherwise, stderr.')
-    parser.add_argument('--threads', '-t', type=int, default=1, help='Number of threads used')
+    parser.add_argument('--gene', '-g', type=str.upper, help='Target gene, like in a capture.')
+    parser.add_argument('--output_dir', '-o', default='.', help='Output directory.')
+    parser.add_argument('--log', '-l', help='Specify log file instead of stderr.')
+    parser.add_argument('--threads', '-t', type=int, default=1, help='Number of threads used.')
+    parser.add_argument('--force_overwrite', '-f', action='store_true',
+                        help='Output results file will be overwritten, but not the alignment.')
     args = parser.parse_args()
+
     # Setup logging
     log_format = '%(asctime)s - %(levelname)s: %(message)s'
     date_format = '%Y/%m/%d %I:%M:%S %p'  # 2010/12/12 11:46:36 AM
@@ -80,17 +78,24 @@ def main():
 
     # Check if final output file already exists
     results_file_name = args.output_dir + '/results.txt'
-    if os.path.exists(results_file_name):
+    if os.path.exists(results_file_name) and not args.force_overwrite:
         logging.error('Final output file already exists. Exiting...')
         sys.exit()
+    if os.path.exists(results_file_name) and args.force_overwrite:
+        logging.warning('Overwriting output results file...')
 
     # Parse Factera output and locate relevant fusions
-    logging.info('Extracting fusions of specified gene pair...')
+    logging.info('Extracting fusions...')
     factera_fusions_reader = csv.DictReader(
         args.factera_fusions_txt, delimiter='\t')
     fusions = dict()
     for index, row_dict in enumerate(factera_fusions_reader):
         row_dict['index'] = index
+        # If a target gene is specified but is not one of the genes involved
+        # in the fusion, skip to the next fusion
+        if args.gene and not (row_dict['Region1'].upper() == args.gene or
+                              row_dict['Region2'].upper() == args.gene):
+            continue
         fusions[index] = row_dict
 
     # Extract sequences for fusions from Factera output
@@ -100,11 +105,14 @@ def main():
     factera_fusionseqs = open(
         factera_outout_prefix + '.factera.fusionseqs.fa', 'r')
     for seq_dict in fasta_gen(factera_fusionseqs):
+        # If ever the FASTA index isn't in the list of fusion indices
+        # (which can happen when a target gene is specified), skip
+        if seq_dict['index'] not in fusions.keys():
+            continue
         fusions[seq_dict['index']]['fusion_name'] = seq_dict['name']
         fusions[seq_dict['index']]['fusion_seq'] = seq_dict['seq']
         fusions[seq_dict['index']]['fusion_ref_name'] = (
-            '{Region1}_{Region2}_{Break1}_{Break2}'.format(**fusions[seq_dict['index']])
-        )
+            '{Region1}_{Region2}_{Break1}_{Break2}'.format(**fusions[seq_dict['index']]))
 
     # Create output directory if doesn't exist
     if not os.path.exists(args.output_dir):
@@ -121,8 +129,10 @@ def main():
     output_sam = args.output_dir + '/bwa_sampe.sam'
     output_bam = args.output_dir + '/bwa_sampe.bam'
     output_bam_sorted = args.output_dir + '/bwa_sampe.sorted.bam'
+    # output_bam_sorted_rmdup = args.output_dir + '/bwa_sampe.sorted.rmdup.bam'
     ## Check if final output BAM file and index exist
-    if os.path.exists(output_bam_sorted) and os.path.exists(output_bam_sorted + '.bai'):
+    if (os.path.exists(output_bam_sorted)
+            and os.path.exists(output_bam_sorted + '.bai')):
         logging.warning('Final output BAM file and index already exist. Skipping alignment...')
     else:
         # Generate new reference genome with fusions
@@ -183,6 +193,12 @@ def main():
         else:
             logging.info('Sorting converted BAM file...')
             pysam.sort('-f', output_bam, output_bam_sorted)
+        # ## Removing duplicates
+        # if os.path.exists(output_bam_sorted_rmdup):
+        #     logging.warning('Duplicates already removed. Skipping...')
+        # else:
+        #     logging.info('Removing duplicates from BAM file...')
+        #     pysam.rmdup(output_bam_sorted, output_bam_sorted_rmdup)
         ## Indexing sorted BAM file
         if os.path.exists(output_bam_sorted + '.bai'):
             logging.warning('Index for sorted BAM file already exists. Skipping...')
@@ -210,16 +226,20 @@ def main():
         bp2_chr, bp2_pos = fusion['Break2'].split(':')
         bp1_cov = calc_cov_at_pos(bam_file, bp1_chr, bp1_pos, args.window, args.min_overlap)
         bp2_cov = calc_cov_at_pos(bam_file, bp2_chr, bp2_pos, args.window, args.min_overlap)
-        wildtype_spanning_reads = bp1_cov[0] + bp2_cov[0]
-        wildtype_spanning_read_pairs = bp1_cov[1] + bp2_cov[1]
+        wildtype_spanning_reads = len(bp1_cov[0]) + len(bp2_cov[0])
+        wildtype_spanning_read_pairs = len(bp1_cov[1]) + len(bp2_cov[1])
         # Calculate support and vaf
         fusion_total_support = fusion_spanning_reads + fusion_spanning_read_pairs
         wildtype_total_support = wildtype_spanning_reads + wildtype_spanning_read_pairs
-        variant_allele_fraction = round(
-            float(fusion_total_support) / float(wildtype_total_support), 1)
+        if wildtype_total_support != 0:
+            # The min() serves to cap the VAF at 1.0, just to be safe
+            variant_allele_fraction = min(round(
+                float(fusion_total_support) / float(wildtype_total_support), 3), 1.0)
+        else:
+            variant_allele_fraction = 'N/A'
         # Compile results
         results.append({
-            'fusion': fusion['fusion_ref_name'],
+            'fusion_id': fusion['fusion_ref_name'],
             'gene_1': fusion['Region1'],
             'gene_2': fusion['Region2'],
             'breakpoint_1': fusion['Break1'],
@@ -236,14 +256,19 @@ def main():
     # Write out results to output file
     logging.info('Outputting results...')
     with open(results_file_name, 'w') as results_file:
-        csv_writer = csv.DictWriter(results_file, OUTPUT_FIELDNAMES)
+        csv_writer = csv.DictWriter(results_file, OUTPUT_FIELDNAMES, delimiter='\t')
+        csv_writer.writeheader()
         csv_writer.writerows(results)
 
     # Clean up only after final output is created
-    os.remove(output_sai_1)
-    os.remove(output_sai_2)
-    os.remove(output_sam)
-    os.remove(output_bam)
+    if os.path.exists(output_sai_1):
+        os.remove(output_sai_1)
+    if os.path.exists(output_sai_2):
+        os.remove(output_sai_2)
+    if os.path.exists(output_sam):
+        os.remove(output_sam)
+    if os.path.exists(output_bam):
+        os.remove(output_bam)
 
 
 def fasta_gen(file_object):
@@ -295,9 +320,15 @@ def calc_cov_at_pos(aln_file, chromosome, pos, window=2000, min_overlap=10):
     spanning_reads = []
     spanning_read_pairs = []
     non_spanning_reads = defaultdict(list)
+    # Ensure that position, window and min_overlap are integers
+    pos = int(pos)
+    window = int(window)
+    min_overlap = int(min_overlap)
     # Extract reads from specified region from alignment file
-    start = pos - window / 2
-    end = pos + window / 2
+    # max used to ensure the start is at least 0
+    start = max(pos - window / 2, 0)
+    # max used to ensure that end is at least 1 bp more than start
+    end = max(pos + window / 2, start + 1)
     for read in aln_file.fetch(chromosome, start, end):
         # Check if read spans breakpoint
         if (read.reference_start < pos - min_overlap and read.reference_end > pos + min_overlap):
@@ -322,6 +353,9 @@ def calc_cov_at_pos(aln_file, chromosome, pos, window=2000, min_overlap=10):
                          pos + min_overlap)):
                     spanning_read_pairs.append((r1, r2))
                     continue
+    # If window is 0, ignore spanning read pairs
+    if window == 0:
+        spanning_read_pairs = []
     return (spanning_reads, spanning_read_pairs)
